WIP code comments
