+++ core/src/constants.rs
@@ -1,19 +1,15 @@
-use bitcoin::BlockHash;
-// use clementine_circuits::constants::CLAIM_MERKLE_TREE_DEPTH;
+use bitcoin::{Amount, BlockHash};
+
 use crypto_bigint::U256;
-
-// /// For deposits, bridge operator does not accept the tx if it is not confirmed
-// pub const CONFIRMATION_BLOCK_COUNT: u32 = 1;
-
-/// K_DEEP is the give time to verifier to make a proper challenge
-pub const K_DEEP: u32 = 3;
-
-/// MAX_BITVM_CHALLENGE_RESPONSE_BLOCKS is maximum number of blocks a single bitvm challenge response can take
-pub const MAX_BITVM_CHALLENGE_RESPONSE_BLOCKS: u32 = 5;
 
 pub type VerifierChallenge = (BlockHash, U256, u8);
 
-// pub const TEST_MODE: bool = true;
-
-// dummy number of BitVM disprove scripts
+/// Dummy number of BitVM disprove scripts
 pub const NUM_INTERMEDIATE_STEPS: usize = 100;
+
+pub const KICKOFF_UTXO_AMOUNT_SATS: Amount = Amount::from_sat(100_000);
+
+pub const KICKOFF_INPUT_AMOUNT: Amount = Amount::from_sat(100_000);
+pub const MIN_TAPROOT_AMOUNT: Amount = Amount::from_sat(330); // TODO: Maybe this could be 294, check
+pub const ANCHOR_AMOUNT: Amount = Amount::from_sat(240); // TODO: This will change to 0 in the future after Bitcoin v0.29.0
+pub const OPERATOR_CHALLENGE_AMOUNT: Amount = Amount::from_sat(200_000_000);
