+++ core/tests/musig2.rs
@@ -4,13 +4,13 @@ use bitcoin::secp256k1::{Message, PublicKey};
 use bitcoin::XOnlyPublicKey;
 use bitcoin::{hashes::Hash, script, Amount, ScriptBuf};
 use bitcoincore_rpc::RpcApi;
+use clementine_core::builder::sighash::{convert_tx_to_pubkey_spend, convert_tx_to_script_spend};
 use clementine_core::builder::transaction::TxHandler;
 use clementine_core::musig2::{
     aggregate_nonces, aggregate_partial_signatures, AggregateFromPublicKeys, Musig2Mode,
 };
 use clementine_core::utils::{handle_taproot_witness_new, SECP};
 use clementine_core::{
-    actor::Actor,
     builder::{self},
     config::BridgeConfig,
     extended_rpc::ExtendedRpc,
@@ -105,7 +105,7 @@ async fn key_spend() {
     };
 
     let message = Message::from_digest(
-        Actor::convert_tx_to_sighash_pubkey_spend(&mut tx_details, 0)
+        convert_tx_to_pubkey_spend(&mut tx_details, 0, None)
             .unwrap()
             .to_byte_array(),
     );
@@ -204,7 +204,7 @@ async fn key_spend_with_script() {
         out_taproot_spend_infos: vec![Some(to_address_spend.clone())],
     };
     let message = Message::from_digest(
-        Actor::convert_tx_to_sighash_pubkey_spend(&mut tx_details, 0)
+        convert_tx_to_pubkey_spend(&mut tx_details, 0, None)
             .unwrap()
             .to_byte_array(),
     );
@@ -310,7 +310,7 @@ async fn script_spend() {
         out_taproot_spend_infos: vec![None],
     };
     let message = Message::from_digest(
-        Actor::convert_tx_to_sighash_script_spend(&mut tx_details, 0, 0)
+        convert_tx_to_script_spend(&mut tx_details, 0, 0, None)
             .unwrap()
             .to_byte_array(),
     );
